Removes CASinoCore::Settings in favor of using ActiveSupport::Configurable
The impetus for this is two-fold.

First, I have some other upcoming changes and I needed some sort of base
configuration to work within. Prior to this change, the
`CASinoCore.setup` call happened too late in the startup process so it
was unavailable. This works around that by always proving a
configuration interface.

Second, ActiveSupport::Configurable already handles most of the same
functionality that CASinoCore::Settings provided. So this offloads the
maintenance responsibility onto ActiveSupport. Open source FTW!

Note: This is a breaking change and will require updates to the
companion CASino Rails Engine gem as well as the online documentation.